fuse expand and parse passes for `compile`
Moving parsing into the expander (that is, making the expander
optionally return a `parsed` record instead of a syntax object)
provides a shotcut for the case that the result of expansion
will be sent on to the compiler. It's a shortcut because the
compiler doesn't have to resolve bindings again or pattern match
on syntax objects.

Module expansion is significantly complicated by having to track
both expander and parsed results at once.