Use "improper lists" in places where single-element lists are common
Some tables support a list of values where a single value is common.
In two cases that seem to be a source of many pair allocations, use
an "improper list" datatype to avoid the pair in most cases.

This change removes 0.5%-1% of the pairs allocated during
`racket -cl racket/base`.