avoid creating instance-access toplevel if unneeded
The runtime "prefix" for a linklet holds all the module-level
variables that are accessed within the linklet. The first slot of the
prefix holds a dummy variable that is used to get access to the
linklet's instance.

Don't create that dummy variable if it isn't needed. That half-step
toward space safety is needed for `db/private/generic/prepared`,
so that the wil-executing thread doesn't hold on to the whole
namespace.

Proper space safety needs a little more, but this improvement
catches back up with the old expander/compiler.