improve linklet flattern's pruning of definitions
Infer pure functions, including some tracking of local variable stypes
so that `(and (list? v) (list-length v))` can be inferred as pure, and
so on.

The goal (not completely achieved) is to eliminate all run-time
support for keyword functions, since they should be expanded away
completely in the expander and regexp implementations.