add a function-lifting pass alongside schemify
This pass is probably temporary, since it would be better done
post-cp0 using the compiler's representation of expressions.

In a case like

 (let loop (....)
   (define (f ...) <use free variables>)
   .... <use f only in application position> ...)

the closure for `f` can be eliminated by adjusting each call to `f` to
pass along the free variables. This is not guaranteed to be faster
than creating a closure, but it's likely to be. The Racket bytecode
compiler reliably performs that transformation, and some Racket
programmers (like me) rely on it.

Chez Scheme doesn't perform the transformation, although inlinig often
eliminates a closure. In the case of `continue` in "utf-8-decode.rkt"
of the "io" implementation, however, `continue` is allocated as a
closure each time around its enclosing loop.

Adding a function liftng pass after Schemify reduces the expander's
allocation significantly loading `racket/base` frmo source, although
the saved allocation doesn't lead to much speedup overall. At the same
time, the transformation itself is fast enough to not cause a slowdown,
so it seems worthwhile for reasoning about allocation.

Putting the lifting pass right after schemify is not ideal, because it
can intefere with constant and copy propagation.