cs: fix weak reference by custodian
Simplify handling of the "stubborn" (i.e., ordered) will executor that
is used by a custodian.

To work correctly with this custodian implementation, a library that
uses `unsafe-custodian-register` will need to register an object that
is not accessible to safe code. (This will be a corner of
incompatibility between conventional Racket and Chez-based Racket.)
Registering an inaccessible object may require an indirection, and
file ports now use that indirection.