add experimental JIT mode to Racket-on-Chez
Experiment with "compiling" linklets to S-expressions, interpreting up
to the point of `lambda`, and calling the Chez Scheme compiler on-demand
when a `lambda` is applied.

The potential benefits are reducing compilation time, reducing ".zo"
file size, and making ".zo" files platform-indepdendent and less
sensitive to changes to the core. The potential drawbacks are slower
run-time performance due to on-demand compilation.

Based on this initial experiment, JIT mode is not obvioulsy better,
but it's promising. Some of the performance penalty is in calling
`interpret` on the code that surrounds `lambda` forms, and a custom,
simple interpreter is likely to help there. Some of the performance
penalty is in making indvidual `lambda` forms separately compiled, so
even less cross-function optimizations happen. That problem might be
addressed by implementing our own optimization front-end (including
cross-linklet optimization) in the schemify level.