cs: avoid unhelpful continuation-mark caching
Cache only if several continuation frames must be consulted to find
the answer.

This change reduces the effect of a minor of space leak in all variants
of Racket: looking for a continuation mark that doesn't have a value
can cause the mark key itself to be retained in the cache (with a "not
there" cached result).