cs: lift lambdas in Rumble implementation
Within the Rumble layer, redefine `define` to lift nested functions
and avoid closure allocation, which provides a more Racket-like
allocation model to the implementation of Rumble.

Lifting doesn't make a big difference in time. Although lifting
everywhere reduces 10% of the total allocation for `make expand-demo`,
it hs a tiny effect on overall run time. Lifting does help with
`equal?` on simple values, reducing all allocation and 10-20% of the
time. It's normally helpful or harmless, so it's the default. Lifting
turns out to be counterproductive for `continuation-mark-set-first`
and `map`, so it's disabled there.