Make Racket-on-Chez report 'cs for `(system-type 'gc)`
In principle, it would be better for `(system-type 'gc)`
to report something like 'conservative or 'precise, and
a separate `(system-type 'variant)` would report 'cgc,
'3m, or 'cs --- although that result could be inferred from
a combination of `(system-type 'cs)` and `(system-type 'vm)`.
Since there will be some kind of redundancy, anyway, it's
simpler and more compatible to let `(system-type 'gc)`
effectively report the Racket variant.