expander: accomodate lost `require`s in nested `#%module-begin`
If the 'module-begin context is used multiple times while expanding a
module, accumulate `require` dependencies across the different
expansions. That way, a nested expansion or a replacement to be
re-expanded can use pieces that were synthesized from an earlier
expansion and that may have bindings to modules not referenced
in the current expansion; that's a questionable practice by a
`#%module-begin` macro, but support it for backward compatibility.